control_plane/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.27.0
pydantic==1.10.14
pyyaml==6.0.1
redis==5.0.3
